Added a whole bunch of stuff.
CardDisplayView -> Tried to add animations but it didn't work yet
ClientEventManager -> Added a parameter for our instance of MainActivity to allow Ui updates
ClientObject -> Added spots to pass in MainActivity as a parameter
ConfigurationActivity -> Moved game creation out of here into MainActivity so that it could be passed to the Event Manager
GameState -> removed instances of the Client and Server event manager, and simply instantiated them once as an object rather than a class.
HolderView -> Changed rotation logic to allow for 90 degree turns only.
LayoutConfigView -> Changed startPositions for loops to how they should've been. Messing with ranges is complicated. Also removed layoutParams from the intent generator and replaced them with the general property getter
MainActivity -> Added logic to manage creation of the game, since that was removed from ConfigurationActivity. Also tried to add animations for flipping but they didn't work.
Pile -> initialized the scale values so that I could try to implement flip animations but again, they didn't work.
ServerEventManager -> Added mainActivity logic identical to ClientEventManager's above
ServerObject -> Added the MainActivity parameter to allow it to pass MainActivity to the ServerEventManager
Splay -> Added a rough draft of splay rotation. It adds logic for handling rotation on initialization and also during setCardPositions. Also added an extra line for the flip method to flip the cards within the splay rather than solely the attached CardViews collection.