Completely Overhauled the Splay class
Changed splay from a list of cards to an imageView that draws its own bitmap.
There are still glitches, and among them are cutting off the cards value at the start of the game.
The onTouchEvent for splay is not implemented yet, but should make things easier when it is.